Rename module features -> capabilities
