+++ packages/vm/src/capabilities.rs
