+++ packages/vm/src/lib.rs
@@ -3,12 +3,12 @@
 mod backend;
 mod cache;
 mod calls;
+mod capabilities;
 mod checksum;
 mod compatibility;
 mod conversion;
 mod environment;
 mod errors;
-mod features;
 mod imports;
 mod instance;
 mod limited;
@@ -37,12 +37,12 @@ pub use crate::calls::{
     call_ibc_packet_ack, call_ibc_packet_ack_raw, call_ibc_packet_receive,
     call_ibc_packet_receive_raw, call_ibc_packet_timeout, call_ibc_packet_timeout_raw,
 };
+pub use crate::capabilities::capabilities_from_csv;
 pub use crate::checksum::Checksum;
 pub use crate::errors::{
     CommunicationError, CommunicationResult, RegionValidationError, RegionValidationResult,
     VmError, VmResult,
 };
-pub use crate::features::capabilities_from_csv;
 pub use crate::instance::{GasReport, Instance, InstanceOptions};
 pub use crate::serde::{from_slice, to_vec};
 pub use crate::size::Size;
